// raahyab/lib/schemas/contact.ts
import { z } from "zod";

export const contactSchema = z.object({
  name: z.string().min(2, "Name must be alteast 2 characters"),
  email: z.string().email("PLease enter a valid email"),
  subject: z.string().min(6, "Subject must be atleast 6 characters"),
  message: z.string().min(10, "Message must be at least 10 characters"),
});

export type ContactFormData = z.infer<typeof contactSchema>;